Refactor data extraction
