+++ app.js
@@ -1,8 +1,9 @@
 const express = require('express');
 require('dotenv').config();
 const routes = require('./routes/api');
-const PORT = process.env.PORT || 5001;
+const PORT = process.env.PORT || 3001;
 const app = express();
+
 app.use('/api', routes);
 
 app.listen(PORT, (err) => {
