+++ controllers/movieGenreController.js
@@ -1,8 +1,12 @@
 const queries = require("../lib/db/queries/index");
 
 const getMovieGenre = async (req, res) => {
-  const data = await queries.getMostPopularGenre(req.query.year);
-  res.status(200).json(data);
+  try {
+    const data = await queries.getMostPopularGenre(req.query.year);
+    res.status(200).json(data);
+  } catch (err) {
+    res.status(500).message('Error: ', err);
+  }
 }
 
 exports.getMovieGenre = getMovieGenre;
